update
on successful logon, redirects to dinosaur page
