+++ client/src/App.jsx
@@ -4,9 +4,9 @@ import HomePage from "./components/Home/HomePage";
 import NewsPage from "./components/News/NewsPage";
 import Footer from "./components/Footer/Footer";
 import DinosaurPage from "./components/Dinosaurs/DinosaurPage";
-import Login from "./components/Login";
-import Register from "./components/Register";
-import Success from "./components/Success";
+import Login from "./components/auth/Login";
+import Register from "./components/auth/Register";
+import Success from "./components/auth/Success";
 import Chart from "./components/Chart/Chart";
 import DinosaurLocation from "./components/Dinosaurs/DinosaurLocation";
 
@@ -20,8 +20,8 @@ function App() {
           <Route path="/" element={<HomePage />} />
           <Route path="/dinosaurs" element={<DinosaurPage />} />
           <Route path="/news" element={<NewsPage />}></Route>
-          <Route path="/Login" element={<Login />} />
-          <Route path="/Register" element={<Register />} />
+          <Route path="/auth/Login" element={<Login />} />
+          <Route path="/auth/Register" element={<Register />} />
           <Route path="/Success" element={<Success />} />
           <Route path="/chart" element={<Chart />} />
           <Route path="/location" element={<DinosaurLocation />} />
