+++ client/src/components/Header/Header.jsx
@@ -1,8 +1,9 @@
-import { useState } from "react";
-import { useNavigate } from "react-router-dom";
-import { NavLink } from "react-router-dom";
+import { useState, useEffect } from "react";
+import { useNavigate, NavLink } from "react-router-dom";
 
-import DinoLogo from "../../assets/dino-logo.png";
+import { isLoggedIn, getUserName } from "../auth/authService";
+
+import DinoLogo from "../../assets/Dinosaur.jpg";
 
 const Header = () => {
   const [isClicked, setIsClicked] = useState(false);
@@ -11,6 +12,16 @@ const Header = () => {
   const toggleBtn = () => {
     setIsClicked(!isClicked);
   };
+
+  const [loggedIn, setLoggedIn] = useState(false);
+  const [userName, setUserName] = useState("");
+
+  useEffect(() => {
+    if (isLoggedIn()) {
+      setLoggedIn(true);
+      setUserName(getUserName());
+    }
+  }, []);
 
   return (
     <header className="p-[15px] md:p-[18px] flex flex-row justify-between items-center text-emerald-100   md:my-0 fixed z-50 top-0 left-0 right-0 bg-gradient-to-tl from-green-950 to-black-900 ">
@@ -30,27 +41,31 @@ const Header = () => {
         </NavLink>
       </nav>
 
-      <div onClick={toggleBtn}>
-        <button className="p-[8px] bg-emerald-500 text-[#fff] font-bold rounded-xl md:text-[20px] relative">
-          Get started
-        </button>
-        {isClicked && (
-          <div className="absolute flex flex-col justify-center items-center bg-[#fff] text-emerald-500 rounded-b-xl rounded-t right-[18px] w-[125px] p-[10px]">
-            <p
-              className="p-[5px] hover:font-bold cursor-pointer"
-              onClick={() => navigate("/register")}
-            >
-              Sign Up
-            </p>
-            <p
-              className="p-[5px] hover:font-bold cursor-pointer"
-              onClick={() => navigate("login")}
-            >
-              Log In
-            </p>
-          </div>
-        )}
-      </div>
+      {loggedIn ? (
+        <p>Welcome, {userName}</p>
+      ) : (
+        <div onClick={toggleBtn}>
+          <button className="p-[8px] bg-emerald-500 text-[#fff] font-bold rounded-xl md:text-[20px] relative">
+            Get started
+          </button>
+          {isClicked && (
+            <div className="absolute flex flex-col justify-center items-center bg-[#fff] text-emerald-500 rounded-b-xl rounded-t right-[18px] w-[125px] p-[10px]">
+              <p
+                className="p-[5px] hover:font-bold cursor-pointer"
+                onClick={() => navigate("/auth/register")}
+              >
+                Sign Up
+              </p>
+              <p
+                className="p-[5px] hover:font-bold cursor-pointer"
+                onClick={() => navigate("/auth/login")}
+              >
+                Log In
+              </p>
+            </div>
+          )}
+        </div>
+      )}
     </header>
   );
 };
